Migrated draw to ebiten

+++ collision/collision.go
@@ -3,22 +3,22 @@ package collision
 type Rectangle struct {
 	X      float64
 	Y      float64
-	Width  float64
-	Height float64
+	Width  int
+	Height int
 }
 
 type Boundary interface {
 	GetX() float64
 	GetY() float64
-	GetWidth() float64
-	GetHeight() float64
+	GetWidth() int
+	GetHeight() int
 }
 
 func CheckCollision(b1 Boundary, b2 Boundary) bool {
-	return b1.GetX() < b2.GetX()+b2.GetWidth() &&
-		b1.GetX()+b1.GetWidth() > b2.GetX() &&
-		b1.GetY() < b2.GetY()+b2.GetHeight() &&
-		b1.GetY()+b1.GetHeight() > b2.GetY()
+	return int(b1.GetX()) < int(b2.GetX())+b2.GetWidth() &&
+		int(b1.GetX())+b1.GetWidth() > int(b2.GetX()) &&
+		int(b1.GetY()) < int(b2.GetY())+b2.GetHeight() &&
+		int(b1.GetY())+b1.GetHeight() > int(b2.GetY())
 }
 
 func (r *Rectangle) GetX() float64 {
@@ -29,10 +29,10 @@ func (r *Rectangle) GetY() float64 {
 	return r.Y
 }
 
-func (r *Rectangle) GetWidth() float64 {
+func (r *Rectangle) GetWidth() int {
 	return r.Width
 }
 
-func (r *Rectangle) GetHeight() float64 {
+func (r *Rectangle) GetHeight() int {
 	return r.Height
 }
